+++ catsquad-shared/src/api/password_change_add.rs
@@ -2,7 +2,7 @@ pub const LINK_API_PASSWORD_CHANGE_ADD: &str = "/api/password_change_add";
 
 #[derive(Clone, Debug, serde::Serialize, serde::Deserialize, PartialEq)]
 pub struct PasswordChangeRes {
-    pub expires: u128,
+    pub expires: u64,
 }
 
 #[derive(Clone, Debug, serde::Serialize, serde::Deserialize, PartialEq)]
